refactor(sp): group sequence-parallel implementation
